fix(auth): address oauth consent review issues
